Add router-core package
