+++ package.json
@@ -74,15 +74,16 @@
       "@tanstack/react-query": "5.64.2",
       "use-sync-external-store": "1.2.2",
       "@tanstack/history": "workspace:*",
+      "@tanstack/router-core": "workspace:*",
       "@tanstack/react-cross-context": "workspace:*",
       "@tanstack/react-router": "workspace:*",
-      "@tanstack/router-cli": "workspace:*",
+      "@tanstack/react-router-with-query": "workspace:*",
       "@tanstack/router-devtools": "workspace:*",
       "@tanstack/router-generator": "workspace:*",
-      "@tanstack/virtual-file-routes": "workspace:*",
+      "@tanstack/router-cli": "workspace:*",
       "@tanstack/router-plugin": "workspace:*",
       "@tanstack/router-vite-plugin": "workspace:*",
-      "@tanstack/react-router-with-query": "workspace:*",
+      "@tanstack/virtual-file-routes": "workspace:*",
       "@tanstack/zod-adapter": "workspace:*",
       "@tanstack/valibot-adapter": "workspace:*",
       "@tanstack/arktype-adapter": "workspace:*",
