+++ packages/react-router/src/routeInfo.ts
@@ -1,8 +1,12 @@
 import type { InferFileRouteTypes } from './fileRoute'
-import type { AddTrailingSlash, RemoveTrailingSlashes } from './link'
+import type {
+  AddTrailingSlash,
+  PartialMergeAll,
+  RemoveTrailingSlashes,
+  TrailingSlashOption,
+} from '@tanstack/router-core'
 import type { AnyRoute } from './route'
-import type { AnyRouter, TrailingSlashOption } from './router'
-import type { PartialMergeAll } from './utils'
+import type { AnyRouter } from './router'
 
 export type ParseRoute<TRouteTree, TAcc = TRouteTree> = TRouteTree extends {
   types: { children: infer TChildren }
