+++ packages/react-router/src/router.ts
@@ -5,27 +5,26 @@ import {
 } from '@tanstack/history'
 import { Store, batch } from '@tanstack/react-store'
 import invariant from 'tiny-invariant'
-import { rootRouteId } from './root'
-import { defaultParseSearch, defaultStringifySearch } from './searchParams'
-import {
-  createControlledPromise,
-  deepEqual,
-  functionalUpdate,
-  last,
-  pick,
-  replaceEqualDeep,
-} from './utils'
 import {
   cleanPath,
+  createControlledPromise,
+  deepEqual,
+  defaultParseSearch,
+  defaultStringifySearch,
+  functionalUpdate,
   interpolatePath,
   joinPaths,
+  last,
   matchPathname,
   parsePathname,
+  pick,
+  replaceEqualDeep,
   resolvePath,
+  rootRouteId,
   trimPath,
   trimPathLeft,
   trimPathRight,
-} from './path'
+} from '@tanstack/router-core'
 import { isRedirect, isResolvedRedirect } from './redirects'
 import { isNotFound } from './not-found'
 import type { StartSerializer } from './serializer'
@@ -37,9 +36,26 @@ import type {
   RouterHistory,
 } from '@tanstack/history'
 import type { NoInfer } from '@tanstack/react-store'
-import type { Manifest } from './manifest'
+
 import type {
   AnyContext,
+  AnySchema,
+  AnyValidator,
+  CommitLocationOptions,
+  ControlledPromise,
+  Manifest,
+  NonNullableUpdater,
+  ParsedLocation,
+  PickAsRequired,
+  ResolveRelativePath,
+  SearchMiddleware,
+  SearchParser,
+  SearchSerializer,
+  TrailingSlashOption,
+  Updater,
+  ViewTransitionOptions,
+} from '@tanstack/router-core'
+import type {
   AnyRoute,
   AnyRouteWithContext,
   BeforeLoadContextOptions,
@@ -50,8 +66,8 @@ import type {
   RouteComponent,
   RouteContextOptions,
   RouteMask,
-  SearchMiddleware,
 } from './route'
+
 import type {
   FullSearchSchema,
   RouteById,
@@ -60,28 +76,17 @@ import type {
   RoutesByPath,
 } from './routeInfo'
 import type {
-  ControlledPromise,
-  NonNullableUpdater,
-  PickAsRequired,
-  Updater,
-} from './utils'
-import type {
   AnyRouteMatch,
   MakeRouteMatch,
   MakeRouteMatchUnion,
   MatchRouteOptions,
 } from './Matches'
-import type { ParsedLocation } from './location'
-import type { SearchParser, SearchSerializer } from './searchParams'
-import type {
-  BuildLocationFn,
-  CommitLocationOptions,
-  NavigateFn,
-} from './RouterProvider'
+
+import type { BuildLocationFn, NavigateFn } from './RouterProvider'
+
 import type { AnyRedirect, ResolvedRedirect } from './redirects'
 import type { NotFoundError } from './not-found'
-import type { NavigateOptions, ResolveRelativePath, ToOptions } from './link'
-import type { AnySchema, AnyValidator } from './validators'
+import type { NavigateOptions, ToOptions } from './link'
 
 declare global {
   interface Window {
@@ -136,8 +141,6 @@ export type RouterContextOptions<TRouteTree extends AnyRoute> =
     : {
         context: InferRouterContext<TRouteTree>
       }
-
-export type TrailingSlashOption = 'always' | 'never' | 'preserve'
 
 export type InjectedHtmlEntry = Promise<string>
 
@@ -496,10 +499,6 @@ export interface BuildNextOptions {
 export interface MatchedRoutesResult {
   matchedRoutes: Array<AnyRoute>
   routeParams: Record<string, string>
-}
-
-export interface ViewTransitionOptions {
-  types: Array<string>
 }
 
 export type RouterConstructorOptions<
@@ -3027,24 +3026,5 @@ export function getInitialRouterState(
     pendingMatches: [],
     cachedMatches: [],
     statusCode: 200,
-  }
-}
-
-export function defaultSerializeError(err: unknown) {
-  if (err instanceof Error) {
-    const obj = {
-      name: err.name,
-      message: err.message,
-    }
-
-    if (process.env.NODE_ENV === 'development') {
-      ;(obj as any).stack = err.stack
-    }
-
-    return obj
-  }
-
-  return {
-    data: err,
   }
 }
