+++ packages/react-router/src/useLoaderDeps.tsx
@@ -5,7 +5,8 @@ import type {
 } from './structuralSharing'
 import type { AnyRouter, RegisteredRouter } from './router'
 import type { RouteById } from './routeInfo'
-import type { Expand, StrictOrFrom } from './utils'
+import type { StrictOrFrom } from './utils'
+import type { Expand } from '@tanstack/router-core'
 
 export interface UseLoaderDepsBaseOptions<
   TRouter extends AnyRouter,
