+++ packages/react-router/src/useMatch.tsx
@@ -8,7 +8,8 @@ import type {
 } from './structuralSharing'
 import type { AnyRouter, RegisteredRouter } from './router'
 import type { MakeRouteMatch, MakeRouteMatchUnion } from './Matches'
-import type { StrictOrFrom, ThrowOrOptional } from './utils'
+import type { StrictOrFrom } from './utils'
+import type { ThrowOrOptional } from '@tanstack/router-core'
 
 export interface UseMatchBaseOptions<
   TRouter extends AnyRouter,
