+++ packages/router-core/eslint.config.js
@@ -1,0 +1,5 @@
+// @ts-check
+
+import rootConfig from '../../eslint.config.js'
+
+export default [...rootConfig]
