+++ packages/router-core/src/manifest.ts
