+++ packages/router-core/src/root.ts
