+++ packages/router-core/src/searchMiddleware.ts
