+++ packages/router-core/src/structuralSharing.ts
@@ -1,0 +1,7 @@
+import type { Constrain } from './utils'
+
+export interface OptionalStructuralSharing<TStructuralSharing, TConstraint> {
+  readonly structuralSharing?:
+    | Constrain<TStructuralSharing, TConstraint>
+    | undefined
+}
