+++ packages/router-core/tests/path.test.ts
