+++ packages/router-core/tests/utils.test.ts
